requirements.txt
py7zr
numpy
Pillow